update inference pipeline and steps
